requirements.txt
bottle==0.12.8
gunicorn==19.3.0
requests==2.7.0
slacker==0.6.2
